+++ backend/service/fms_api/create_fms_storage_provider_logic.v
@@ -27,8 +27,12 @@ pub fn create_fms_storage_provider_usecase(mut ctx Context, req CreateFmsStorage
 }
 
 fn create_fms_storage_provider_domain(req CreateFmsStorageProviderReq) ! {
-	if req.name == '' { return error('provider name is required') }
-	if req.bucket == '' { return error('bucket is required') }
+	if req.name == '' {
+		return error('provider name is required')
+	}
+	if req.bucket == '' {
+		return error('bucket is required')
+	}
 	if req.secret_id == '' {
 		return error('secret id is required')
 	}
@@ -44,17 +48,17 @@ fn create_fms_storage_provider_domain(req CreateFmsStorageProviderReq) ! {
 }
 
 pub struct CreateFmsStorageProviderReq {
-	name       string  @[json: 'name']
-	bucket     string  @[json: 'bucket']
-	secret_id  string  @[json: 'secretId']
-	secret_key string  @[json: 'secretKey']
-	endpoint   string  @[json: 'endpoint']
+	name       string @[json: 'name']
+	bucket     string @[json: 'bucket']
+	secret_id  string @[json: 'secretId']
+	secret_key string @[json: 'secretKey']
+	endpoint   string @[json: 'endpoint']
 	folder     ?string @[json: 'folder']
-	region     string  @[json: 'region']
-	is_default i16     @[json: 'isDefault']
-	use_cdn    i16     @[json: 'useCdn']
+	region     string @[json: 'region']
+	is_default u8 @[json: 'isDefault']
+	use_cdn    u8 @[json: 'useCdn']
 	cdn_url    ?string @[json: 'cdnUrl']
-	status     i16     @[json: 'status']
+	status     u8 @[json: 'status']
 }
 
 pub struct CreateFmsStorageProviderResp {
@@ -64,18 +68,18 @@ pub struct CreateFmsStorageProviderResp {
 fn create_fms_storage_provider_repo(mut ctx Context, req CreateFmsStorageProviderReq) !CreateFmsStorageProviderResp {
 	time_now := time.now()
 	p := FmsStorageProvider{
-		id:         rand.uuid_v7()
-		name:       req.name
-		bucket:     req.bucket
-		secret_id:  req.secret_id
+		id: rand.uuid_v7()
+		name: req.name
+		bucket: req.bucket
+		secret_id: req.secret_id
 		secret_key: req.secret_key
-		endpoint:   req.endpoint
-		folder:     req.folder
-		region:     req.region
+		endpoint: req.endpoint
+		folder: req.folder
+		region: req.region
 		is_default: req.is_default
-		use_cdn:    req.use_cdn
-		cdn_url:    req.cdn_url
-		status:     req.status
+		use_cdn: req.use_cdn
+		cdn_url: req.cdn_url
+		status: req.status
 		creator_id: ctx.svc_iam.user_id
 		updater_id: ctx.svc_iam.user_id
 		created_at: time_now
